feat: sccSearchOne 잔량 안나오는 현상 수정

+++ src/main/java/com/posco/epro4/DTO/Scc/SccSearchOneDTO.java
@@ -26,5 +26,5 @@ public class SccSearchOneDTO {
     
     private Integer vendor_id;
     private String  vendor_name;
-    private Long    remaining;
+    private Integer    remaining;
 }
